requirements.txt
pyrogram==2.0.37
tgcrypto
aiohttp
python-dotenv
pillow
tqdm
